add final version of bachelor work

+++ src/main/java/com/shadoww/BookLibraryApp/services/BookCatalogsService.java
@@ -7,6 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import jakarta.persistence.EntityNotFoundException;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -14,39 +16,77 @@ import java.util.Optional;
 @Transactional(readOnly = true)
 public class BookCatalogsService {
 
-    private BookCatalogsRepository bookCatalogsRepository;
+    private final BookCatalogsRepository bookCatalogsRepository;
 
     @Autowired
     public BookCatalogsService(BookCatalogsRepository bookCatalogsRepository) {
         this.bookCatalogsRepository = bookCatalogsRepository;
     }
 
-
-    public Optional<BookCatalog> findById(int id) {
-        return bookCatalogsRepository.findById(id);
+    public BookCatalog getByIdAndPerson(int id, Person owner) {
+        return bookCatalogsRepository
+                .findBookCatalogByIdAndOwner(id, owner)
+                .orElseThrow(() -> new EntityNotFoundException("Такого каталога не існує"));
     }
 
-    public Optional<BookCatalog> findByIdAndPerson(int id, Person person) {
-        return bookCatalogsRepository.findBookCatalogByIdAndPerson(id, person);
+    public BookCatalog readById(int id) {
+        return bookCatalogsRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Такого каталога не існує"));
     }
 
-    public List<BookCatalog> findByPerson(Person person) {
 
-        return bookCatalogsRepository.findBookCatalogByPerson(person);
+    public List<BookCatalog> findByPerson(Person owner) {
+
+        return bookCatalogsRepository.findBookCatalogByOwner(owner);
     }
 
     @Transactional
-    public void save(BookCatalog bookCatalog) {
+    public void create(BookCatalog bookCatalog) {
+
+        if (bookCatalog == null) {
+            throw new NullPointerException("Каталог не може бути пустим");
+        }
+
+
         bookCatalogsRepository.save(bookCatalog);
     }
 
+    @Transactional
+    public BookCatalog update(BookCatalog updatedCatalog) {
+
+        if (updatedCatalog == null) {
+            throw new NullPointerException("Каталог не може бути пустим");
+        }
+
+        readById(updatedCatalog.getId());
+
+        return save(updatedCatalog);
+
+    }
+
+    private BookCatalog save(BookCatalog catalog) {
+        if (catalog == null) {
+            throw new NullPointerException("Каталог не може бути пустим");
+        }
+
+        return bookCatalogsRepository.save(catalog);
+    }
+
+    @Transactional
+    public void deleteById(int id) {
+        BookCatalog catalog = readById(id);
+
+        delete(catalog);
+    }
 
     @Transactional
     public void delete(BookCatalog bookCatalog) {
+
+        if (bookCatalog == null) {
+            throw new NullPointerException("Каталог не може бути пустим");
+        }
+
         bookCatalogsRepository.delete(bookCatalog);
     }
-    @Transactional
-    public void deleteCatalog(int id) {
-        bookCatalogsRepository.deleteById(id);
-    }
+
+
 }
